Release derivatives trading option client version 1.1.1
